% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeltIndex-data.R
\docType{data}
\name{MeltIndex}
\alias{MeltIndex}
\title{Melt index of a polyethylene compound}
\format{An object of class xts (inherits from zoo) with 80 rows and 1 columns.}
\source{
Wadsworth, Harrison M., Kenneth S. Stephens, and A. Blanton Godfrey. "Modern methods for quality control and improvement". John Wiley & Sons, 2002.
}
\usage{
data(MeltIndex)
}
\description{
Data set was used  to demonstrate the stepwise approach for setting up a robust Shewhart location and dispersion control charts.
}
\examples{
data(MeltIndex)
}
\references{
Nazir, Hafiz Z., Marit Schoonhoven, Muhammad Riaz, and Ronald JMM Does. "Quality Quandaries: A Stepwise Approach for Setting Up a Robust Shewhart Location Control Chart." Quality Engineering 26, no. 2 (2014): 246-252.

Nazir, Hafiz Z., Marit Schoonhoven, Muhammad Riaz, and Ronald JMM Does. "Quality Quandaries: How to Set Up a Robust Shewhart Control Chart for Dispersion?." Quality Engineering 26, no. 1 (2014): 130-136.
}
\keyword{datasets}
