% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MomentsinPartitions.R
\name{MomentsInPartitions}
\alias{MomentsInPartitions}
\title{The first four moments (mean, standard deviation, skewness, kurtosis) in total, common and special cause periods.}
\usage{
MomentsInPartitions(PartitonedData, Subgroups)
}
\arguments{
\item{PartitonedData}{common and special cause partitions}

\item{Subgroups}{a list of subgroups used for partitioning}
}
\value{
Mean, standard deviation, skewness and kurtosis  in total, common and special caused periods
}
\description{
This function gives the first four moments in total, common and special cause periods. Partitioning of common and special cause periods is based on the Shewhart S chart.
}
\examples{
# subgroup size
SubgroupCriteria="weeks"
# data loading
data(AAPL)
#subgrouping
require(zoo)
StockPrice=window(AAPL[,"Close"],start=as.Date("2012-01-02"),end=as.Date("2014-12-31"))
Subgroups=Subgrouping(StockPrice,SubgroupCriteria,CountSgps=1)
# obtain subgroup summary
DataSum=SubgroupSummary(Subgroups,MaxSgpSize=5)
# partitioning data into common and special cause periods
PartitonedData=PartitioningSgps(DataSum,ChartType="S")
# calculate the first four moments
MomentsInPartitions(PartitonedData$SgpSummary,Subgroups)

}
\references{
Nadeeka Premarathna,  A. Jonathan R. Godfrey and K. Govindaraju. "Decomposition of stock market trade-offs using Shewhart methodology." International Journal of Quality & Reliability Management 33, no. 9 (2016): 1311-1331.
}
\author{
Nadeeka Premarathna
}
