% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MuStepwiseRobustPhaseI.R
\name{MuStepwiseRobustPhaseI}
\alias{MuStepwiseRobustPhaseI}
\title{Unbiased robust estimator for \eqn{\mu} based on a stepwise robust chart procedure}
\usage{
MuStepwiseRobustPhaseI(Subgroups, sigma)
}
\arguments{
\item{Subgroups}{subgroups of the data series, function handles the unequal subgroup sizes.}

\item{sigma}{estimate for \eqn{\sigma}}
}
\value{
estimates of \eqn{\mu} from Phase I data
}
\description{
Estimation procedure combines the use of individual observations and subgroup screening. An initial estimate for \eqn{\mu} is based on the trimmed means of the tri means and screen the subgroups. Then, the resulting estimator for \eqn{\mu} from the remaining subgroups is sued to  screen the individual outliers in the remaining subgroups.
}
\examples{
data(MeltIndex)
require(xts)
Subgroups=split(MeltIndex, f="weeks")
Subgroups[20]=NULL
sigma=SigmaStepwiseRobustPhaseI(Subgroups)
Subgroups=split(MeltIndex, f="weeks")
Subgroups[20]=NULL
MuStepwiseRobustPhaseI(Subgroups,sigma)
}
\references{
Nazir, Hafiz Z., Marit Schoonhoven, Muhammad Riaz, and Ronald JMM Does. "Quality Quandaries: A Stepwise Approach for Setting Up a Robust Shewhart Location Control Chart." Quality Engineering 26, no. 2 (2014): 246-252.
}
\author{
Nadeeka Premarathna
}
