% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartitioningSgps.R
\name{PartitioningSgps}
\alias{PartitioningSgps}
\title{Subgroup partition into common and special causes based on the Shewhart S chart rules}
\usage{
PartitioningSgps(Data, ChartType = c("Xbar", "R", "S"))
}
\arguments{
\item{Data}{subgroup summary (object type "SgpSummary")}

\item{ChartType}{Chart type for partitioning rule: if "Xbar"- Shewhart X-bar chart (mean), "R"- Shewhart control chart for range, "S"- Shewhart S chart}
}
\value{
Subgroups in common and special caused periods
}
\description{
' Function separates subgroups into common and special cause periods.
}
\examples{
SubgroupCriteria="weeks"
data(AAPL)
require(zoo)
StockPrice=window(AAPL[,"Close"],start=as.Date("2012-01-02"),end=as.Date("2013-12-31"))
Subgroups=Subgrouping(StockPrice,SubgroupCriteria,CountSgps=1)
DataSum=SubgroupSummary(Subgroups,MaxSgpSize=5)
B=PartitioningSgps(DataSum,ChartType="S")
}
\references{
Nadeeka Premarathna,  A. Jonathan R. Godfrey and K. Govindaraju. "Decomposition of stock market trade-offs using Shewhart methodology." International Journal of Quality & Reliability Management 33, no. 9 (2016): 1311-1331.
}
\seealso{
\code{\link{Subgrouping}}, \code{\link{SubgroupSummary}}
}
\author{
Nadeeka Premarathna
}
