% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotShewhartCharts.R
\name{ShewhartSgpCharts}
\alias{ShewhartSgpCharts}
\title{Plotting the Shewhart charts for Phase I and Phase II data}
\usage{
ShewhartSgpCharts(Data, PhaseIIData, ChartType, plot = TRUE)
}
\arguments{
\item{Data}{Phase I partitioned  subgroup summary (object type "Partition.SgpSummary")}

\item{PhaseIIData}{Phase II subgroup summary (object type "SgpSummary")}

\item{ChartType}{Chart type "Xbar"- Shewhart X-bar chart, "R"- Shewhart control chart for range, "S"- Shewhart S chart}

\item{plot}{logical. If TRUE the corresponding Shewhart chart is plotted.}
}
\value{
plot Shewhart charts
}
\description{
The function plots the Shewhart charts for subgroups, in Phase I and Phase II.
}
\examples{
require(zoo)
SubgroupCriteria="weeks"
# data loading
data(AAPL)
#subgrouping Phase I data
StockPrice=window(AAPL[,"Close"],start=as.Date("2012-01-02"),end=as.Date("2013-12-31"))
Subgroups=Subgrouping(StockPrice,SubgroupCriteria,CountSgps=1)
# obtain subgroup summary of Phase I data
DataSum=SubgroupSummary(Subgroups,MaxSgpSize=5)
# partitioning data into common and special cause periods (Phase I data)
PartitonedData=PartitioningSgps(DataSum,ChartType="S")

# Phase II data
StockPrice=window(AAPL[,"Close"],start=as.Date("2014-01-06"),end=as.Date("2014-05-30"))
SubgroupsPhaseII=Subgrouping(StockPrice,SubgroupCriteria,CountSgps=1)
PhaseIIData=SubgroupSummary(SubgroupsPhaseII,MaxSgpSize=5)
B=ShewhartSgpCharts(PartitonedData, PhaseIIData,ChartType="S",plot=TRUE)
}
\references{
Nadeeka Premarathna,  A. Jonathan R. Godfrey and K. Govindaraju. "Decomposition of stock market trade-offs using Shewhart methodology." International Journal of Quality & Reliability Management 33, no. 9 (2016): 1311-1331.
}
\author{
Nadeeka Premarathna
}
