% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigmaStepwiseRobustPhaseI.R
\name{SigmaStepwiseRobustPhaseI}
\alias{SigmaStepwiseRobustPhaseI}
\title{Estimating \eqn{\sigma} using a stepwise robust chart procedure}
\usage{
SigmaStepwiseRobustPhaseI(Subgroups)
}
\arguments{
\item{Subgroups}{a list subgroups, function handles the unequal subgroup sizes.}
}
\value{
estimates of \eqn{\sigma} from Phase I data
}
\description{
Estimation procedure performs individual observations and subgroup screening.  An initial estimate for \eqn{\sigma} is obtained from trimmed means of interquartile ranges. Subgroups are first screened.  Then, the resulting \eqn{\sigma} from the remaining subgroups is used to eliminate individual outliers in the remaining subgroups.
}
\examples{
data(MeltIndex)
require(xts)
Subgroups=split(MeltIndex, f="weeks")
Subgroups[20]=NULL
sigma=SigmaStepwiseRobustPhaseI(Subgroups)
}
\references{
Nazir, Hafiz Z., Marit Schoonhoven, Muhammad Riaz, and Ronald JMM Does. "Quality Quandaries: How to Set Up a Robust Shewhart Control Chart for Dispersion?." Quality Engineering 26, no. 1 (2014): 130-136.
}
\author{
Nadeeka Premarathna
}
