% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subgrouping.R
\name{Subgrouping}
\alias{Subgrouping}
\title{Divide into subgroups}
\usage{
Subgrouping(Data, SubgroupCriteria, CountSgps)
}
\arguments{
\item{Data}{a series of time series data (data type: zoo or xts)}

\item{SubgroupCriteria}{length of the subgroups  based on time, eg. "weeks", "months",  if using daliy data}

\item{CountSgps}{subgroups sizes further increase using CountSgps=2,3,4 eg. SubgroupCriteria="weeks", CountSgps=2 gives subgroups size of two weeks}
}
\value{
Subgropus
}
\description{
Function divides the data series into subgroups according to the given criteria.
}
\examples{
SubgroupCriteria="weeks"
data(AAPL)
require(zoo)
StockPrice=window(AAPL[,"Close"],start=as.Date("2012-01-02"),end=as.Date("2014-12-31"))
Subgroups=Subgrouping(StockPrice,SubgroupCriteria,CountSgps=1)
}
\seealso{
\code{\link{PartitioningSgps}}
}
\author{
Nadeeka Premarathna
}
