% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubgroupSummary.R
\name{SubgroupSummary}
\alias{SubgroupSummary}
\title{Subgroup summary}
\usage{
SubgroupSummary(Subgroups, MaxSgpSize)
}
\arguments{
\item{Subgroups}{a series of subgroups}

\item{MaxSgpSize}{possible maximum size of a subgroup}
}
\value{
SgpSummary
}
\description{
Function gives subgroup size, mean, standard deviation, adjusted standard deviation, skewness, kurtosis and a table of control chart constants for subgroups in the given series.
}
\examples{
# Subgroup size
SubgroupCriteria="weeks"
#  load data
data(AAPL)
# divide into subgroups
StockPrice=window(AAPL[,"Close"],start=as.Date("2012-01-02"),end=as.Date("2014-12-31"))
Subgroups=Subgrouping(StockPrice,SubgroupCriteria,CountSgps=1)
# obtain the subgroup summary
SubgroupSummary(Subgroups,MaxSgpSize=5)
}
\seealso{
\code{\link{Subgrouping}}

\code{\link{Subgrouping}}
}
\author{
Nadeeka Premarathna
}
